requirements.txt
aiofiles==23.1.0
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
base58==2.1.1
bitarray==2.8.5
boto3==1.26.165
botocore==1.29.165
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.2.0
cycler==0.12.1
cytoolz==0.12.2
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-typing==3.5.2
eth-utils==2.3.1
fastdtw==0.3.4
fonttools==4.46.0
frozenlist==1.4.0
future==0.18.3
hexbytes==0.3.1
idna==3.4
ipfshttpclient==0.7.0
iso8601==1.1.0
japanize-matplotlib==1.1.3
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lru-dict==1.2.0
matplotlib==3.8.2
minio==7.1.15
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.9.0
numpy==1.24.3
packaging==23.2
parsimonious==0.9.0
Pillow==10.1.0
protobuf==4.25.1
pycryptodome==3.19.0
pyhackrf==0.2.0
pynmea2==1.19.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pyunormalize==15.1.0
PyYAML==6.0
referencing==0.32.0
regex==2023.10.3
requests==2.29.0
rlp==4.0.0
rpds-py==0.13.2
s3transfer==0.6.1
scipy==1.11.1
six==1.16.0
toolz==0.12.0
typing_extensions==4.9.0
urllib3==1.26.15
varint==1.0.2
web3==6.12.0
websockets==12.0
yarl==1.9.4
